actix_security::http::security

Trait SecurityExt

pub trait SecurityExt {
    // Required methods
    fn get_user(&self) -> Option<User>;
    fn is_authenticated(&self) -> bool;
    fn has_role(&self, role: &str) -> bool;
    fn has_any_role(&self, roles: &[&str]) -> bool;
    fn has_authority(&self, authority: &str) -> bool;
    fn has_any_authority(&self, authorities: &[&str]) -> bool;
}

Extension trait for HttpRequest to check authentication.

Required Methods

fn get_user(&self) -> Option<User>

Returns a clone of the authenticated user if present.

fn is_authenticated(&self) -> bool

Returns true if a user is authenticated.

fn has_role(&self, role: &str) -> bool

Checks if the authenticated user has the specified role.

fn has_any_role(&self, roles: &[&str]) -> bool

Checks if the authenticated user has any of the specified roles.

fn has_authority(&self, authority: &str) -> bool

Checks if the authenticated user has the specified authority.

fn has_any_authority(&self, authorities: &[&str]) -> bool

Checks if the authenticated user has any of the specified authorities.